requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
flask_sqlalchemy==3.0.3
passlib==1.7.4
PyJWT==2.6.0
python-dotenv==1.0.0
